test(anthropic-direct): split loop.test.ts into sibling suites (#370)
