+++ src/agent/providers/anthropic-direct/loop.test.ts
@@ -14,197 +14,23 @@
  *  - summarizeToolInput helper behavior
  */
 
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { describe, it, expect, vi, beforeEach } from 'vitest';
 import type { RawMessageStreamEvent, MessageParam, ContentBlockParam } from '@anthropic-ai/sdk/resources';
-import { runTurn, DEFAULT_MAX_TOOL_USE_ITERATIONS, isTransientServerError, isOverloadedErrorEvent, OVERLOAD_MAX_RETRIES } from './loop.js';
+import { runTurn, DEFAULT_MAX_TOOL_USE_ITERATIONS } from './loop.js';
 import type { ProviderEvent } from '../../provider.js';
-import type {
-  AnthropicClientLike,
-  ToolDispatcherLike,
-  ToolCall,
-  ToolResult,
-  TranslateCtx,
-} from './types.js';
-
-// --- Helpers ---
-
-async function* fromArray<T>(arr: T[]): AsyncIterable<T> {
-  for (const x of arr) yield x;
-}
-
-async function collect(gen: AsyncGenerator<ProviderEvent>): Promise<ProviderEvent[]> {
-  const out: ProviderEvent[] = [];
-  for await (const ev of gen) out.push(ev);
-  return out;
-}
-
-const SESSION_ID = 'sess_loop_test';
-const ctx: TranslateCtx = { sessionId: SESSION_ID };
-
-function baseUsage(): {
-  input_tokens: number;
-  output_tokens: number;
-  cache_creation_input_tokens: number;
-  cache_read_input_tokens: number;
-} {
-  return {
-    input_tokens: 10,
-    output_tokens: 5,
-    cache_creation_input_tokens: 0,
-    cache_read_input_tokens: 0,
-  };
-}
-
-function makeTextStream(text: string, stopReason: string = 'end_turn'): RawMessageStreamEvent[] {
-  return [
-    {
-      type: 'message_start',
-      message: {
-        id: 'msg_test',
-        type: 'message',
-        role: 'assistant',
-        content: [],
-        model: 'claude-test',
-        stop_reason: null,
-        stop_sequence: null,
-        usage: baseUsage(),
-      },
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'content_block_start',
-      index: 0,
-      content_block: { type: 'text', text: '' },
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'content_block_delta',
-      index: 0,
-      delta: { type: 'text_delta', text },
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'content_block_stop',
-      index: 0,
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'message_delta',
-      delta: { stop_reason: stopReason, stop_sequence: null },
-      usage: { output_tokens: 5 },
-    } as unknown as RawMessageStreamEvent,
-    { type: 'message_stop' } as unknown as RawMessageStreamEvent,
-  ];
-}
-
-function makeToolUseStream(
-  toolId: string,
-  toolName: string,
-  inputJson: string,
-): RawMessageStreamEvent[] {
-  return [
-    {
-      type: 'message_start',
-      message: {
-        id: 'msg_t',
-        type: 'message',
-        role: 'assistant',
-        content: [],
-        model: 'claude-test',
-        stop_reason: null,
-        stop_sequence: null,
-        usage: baseUsage(),
-      },
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'content_block_start',
-      index: 0,
-      content_block: { type: 'tool_use', id: toolId, name: toolName, input: {} },
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'content_block_delta',
-      index: 0,
-      delta: { type: 'input_json_delta', partial_json: inputJson },
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'content_block_stop',
-      index: 0,
-    } as unknown as RawMessageStreamEvent,
-    {
-      type: 'message_delta',
-      delta: { stop_reason: 'tool_use', stop_sequence: null },
-      usage: { output_tokens: 9 },
-    } as unknown as RawMessageStreamEvent,
-    { type: 'message_stop' } as unknown as RawMessageStreamEvent,
-  ];
-}
-
-function makeMultiToolUseStream(
-  tools: Array<{ id: string; name: string; input: string }>,
-): RawMessageStreamEvent[] {
-  const events: RawMessageStreamEvent[] = [
-    {
-      type: 'message_start',
-      message: {
-        id: 'msg_multi',
-        type: 'message',
-        role: 'assistant',
-        content: [],
-        model: 'claude-test',
-        stop_reason: null,
-        stop_sequence: null,
-        usage: baseUsage(),
-      },
-    } as unknown as RawMessageStreamEvent,
-  ];
-
-  tools.forEach((tool, idx) => {
-    events.push({
-      type: 'content_block_start',
-      index: idx,
-      content_block: { type: 'tool_use', id: tool.id, name: tool.name, input: {} },
-    } as unknown as RawMessageStreamEvent);
-    events.push({
-      type: 'content_block_delta',
-      index: idx,
-      delta: { type: 'input_json_delta', partial_json: tool.input },
-    } as unknown as RawMessageStreamEvent);
-    events.push({
-      type: 'content_block_stop',
-      index: idx,
-    } as unknown as RawMessageStreamEvent);
-  });
-
-  events.push({
-    type: 'message_delta',
-    delta: { stop_reason: 'tool_use', stop_sequence: null },
-    usage: { output_tokens: 10 },
-  } as unknown as RawMessageStreamEvent);
-  events.push({ type: 'message_stop' } as unknown as RawMessageStreamEvent);
-
-  return events;
-}
-
-function makeClient(
-  streamFactory: () => AsyncIterable<RawMessageStreamEvent>,
-): AnthropicClientLike {
-  return {
-    messages: {
-      create: vi.fn(() => streamFactory()),
-    },
-  };
-}
-
-function makeDispatcher(
-  executeFn: (call: ToolCall) => Promise<ToolResult>,
-): ToolDispatcherLike {
-  return { execute: executeFn };
-}
-
-function makeBatchDispatcher(
-  executeBatchFn: (calls: ToolCall[]) => Promise<ToolResult[]>,
-): ToolDispatcherLike {
-  return {
-    execute: vi.fn(() => Promise.reject(new Error('should use batch'))),
-    executeBatch: executeBatchFn,
-  };
-}
+import type { AnthropicClientLike, ToolCall, ToolResult } from './types.js';
+import {
+  fromArray,
+  collect,
+  ctx,
+  baseUsage,
+  makeTextStream,
+  makeToolUseStream,
+  makeMultiToolUseStream,
+  makeClient,
+  makeDispatcher,
+  makeBatchDispatcher,
+} from './loop.test-helpers.js';
 
 // --- Tests ---
 
@@ -1290,1337 +1116,6 @@ describe('loop.ts runTurn', () => {
       // The skill `arguments` field is intentionally NOT echoed — the name
       // alone identifies which skill ran, matching the Agent(label) form.
       expect(loopEmittedStart.toolInput).not.toContain('flaky test');
-    }
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Witness-layer trace emission (PR #2 commit 2)
-// ---------------------------------------------------------------------------
-
-describe('loop.ts runTurn — witness-layer tool_call emission', () => {
-  // Import dynamically inside the describe to avoid forcing an eager
-  // import at module top when the rest of the file does not need it.
-  it('emits tool_call.started before dispatch and tool_call.completed after', async () => {
-    const { InMemoryTraceWriter } = await import('../../trace/writer.js');
-    const writer = new InMemoryTraceWriter();
-
-    // Two streams: first emits a tool_use, second emits final text.
-    let callIdx = 0;
-    const streams = [
-      () => fromArray(makeToolUseStream('tu_1', 'search', '{"q":"hello"}')),
-      () => fromArray(makeTextStream('done', 'end_turn')),
-    ];
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => streams[callIdx++ % streams.length]!()),
-      },
-    };
-    const dispatcher = makeDispatcher(async () => ({
-      content: 'search result',
-      isError: false,
-    }));
-    const messages: MessageParam[] = [{ role: 'user', content: 'search please' }];
-    const ac = new AbortController();
-
-    await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: [{ name: 'search', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: ac.signal,
-        ctx,
-        traceWriter: writer,
-      }),
-    );
-
-    // Drain microtasks so the fire-and-forget emit calls settle.
-    await new Promise((resolve) => setImmediate(resolve));
-
-    const toolCalls = writer.events.filter((e) => e.kind === 'tool_call');
-    expect(toolCalls).toHaveLength(2);
-    if (toolCalls[0]?.kind !== 'tool_call' || toolCalls[1]?.kind !== 'tool_call') {
-      throw new Error('unreachable');
-    }
-    expect(toolCalls[0].payload.phase).toBe('started');
-    expect(toolCalls[1].payload.phase).toBe('completed');
-    if (toolCalls[0].payload.phase !== 'started') throw new Error('unreachable');
-    if (toolCalls[1].payload.phase !== 'completed') throw new Error('unreachable');
-    expect(toolCalls[0].payload.toolUseId).toBe('tu_1');
-    expect(toolCalls[0].payload.name).toBe('search');
-    expect(toolCalls[0].payload.inputBytes).toBeGreaterThan(0);
-    expect(toolCalls[1].payload.toolUseId).toBe('tu_1');
-    expect(toolCalls[1].payload.name).toBe('search');
-    expect(toolCalls[1].payload.resultBytes).toBe(Buffer.byteLength('search result'));
-    expect(toolCalls[1].payload.isError).toBe(false);
-    expect(toolCalls[1].payload.truncated).toBe(false);
-    expect(toolCalls[1].payload.durationMs).toBeGreaterThanOrEqual(0);
-  });
-
-  it('emits exactly one model_ttfb session_phase on the first streamed event', async () => {
-    const { InMemoryTraceWriter } = await import('../../trace/writer.js');
-    const writer = new InMemoryTraceWriter();
-
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => fromArray(makeTextStream('hi', 'end_turn'))),
-      },
-    };
-    const messages: MessageParam[] = [{ role: 'user', content: 'hello' }];
-    const ac = new AbortController();
-
-    await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: [],
-        toolDispatcher: makeDispatcher(async () => ({ content: '', isError: false })),
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: ac.signal,
-        ctx,
-        traceWriter: writer,
-      }),
-    );
-    await new Promise((resolve) => setImmediate(resolve));
-
-    const ttfb = writer.events.filter(
-      (e) =>
-        e.kind === 'session_phase' &&
-        (e.payload as { phase: string }).phase === 'model_ttfb',
-    );
-    // Single text-stream turn => exactly one model API call => one TTFB.
-    expect(ttfb).toHaveLength(1);
-    const payload = ttfb[0]!.payload as { durationMs?: number; resolvedModel?: string };
-    expect(payload.durationMs).toBeTypeOf('number');
-    expect(payload.durationMs).toBeGreaterThanOrEqual(0);
-    // model_ttfb carries the resolved wire id for THIS call (input.model).
-    expect(payload.resolvedModel).toBe('claude-test');
-  });
-
-  it('records isError=true and truncated=true based on tool result content', async () => {
-    const { InMemoryTraceWriter } = await import('../../trace/writer.js');
-    const writer = new InMemoryTraceWriter();
-
-    let callIdx = 0;
-    const streams = [
-      () => fromArray(makeToolUseStream('tu_err', 'bash', '{}')),
-      () => fromArray(makeTextStream('done', 'end_turn')),
-    ];
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => streams[callIdx++ % streams.length]!()),
-      },
-    };
-    const dispatcher = makeDispatcher(async () => ({
-      content: 'partial...[output truncated — exceeded 100KB]',
-      isError: true,
-    }));
-
-    await collect(
-      runTurn({
-        client,
-        messages: [{ role: 'user', content: 'run' }],
-        system: null,
-        tools: [{ name: 'bash', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: new AbortController().signal,
-        ctx,
-        traceWriter: writer,
-      }),
-    );
-
-    await new Promise((resolve) => setImmediate(resolve));
-
-    const completed = writer.events.find(
-      (e) => e.kind === 'tool_call' && e.payload.phase === 'completed',
-    );
-    expect(completed?.kind).toBe('tool_call');
-    if (completed?.kind !== 'tool_call' || completed.payload.phase !== 'completed') {
-      throw new Error('unreachable');
-    }
-    expect(completed.payload.isError).toBe(true);
-    expect(completed.payload.truncated).toBe(true);
-  });
-
-  // Regression: prefer the structured `ToolResult.truncated` flag over the
-  // in-band sentinel string. Before this flag existed, the trace writer
-  // sniffed `result.content.includes('[output truncated')` — fragile
-  // because it relied on a particular sentinel literal staying in content
-  // and because external (third-party) handlers had no clean way to mark
-  // overflow. The dispatcher in this test returns truncated:true with NO
-  // sentinel in content; both the trace and the `tool.output` event must
-  // still reflect truncation.
-  it('propagates ToolResult.truncated=true to trace and tool.output event without sentinel', async () => {
-    const { InMemoryTraceWriter } = await import('../../trace/writer.js');
-    const writer = new InMemoryTraceWriter();
-
-    let callIdx = 0;
-    const streams = [
-      () => fromArray(makeToolUseStream('tu_t', 'bash', '{}')),
-      () => fromArray(makeTextStream('done', 'end_turn')),
-    ];
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => streams[callIdx++ % streams.length]!()),
-      },
-    };
-    // Notably: content has no `[output truncated …]` sentinel string.
-    // Detection must come from the structured flag alone.
-    const dispatcher = makeDispatcher(async () => ({
-      content: 'some legitimate-looking output with no sentinel',
-      truncated: true,
-    }));
-
-    const events = await collect(
-      runTurn({
-        client,
-        messages: [{ role: 'user', content: 'run' }],
-        system: null,
-        tools: [{ name: 'bash', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: new AbortController().signal,
-        ctx,
-        traceWriter: writer,
-      }),
-    );
-
-    await new Promise((resolve) => setImmediate(resolve));
-
-    // Trace records the structured flag.
-    const completed = writer.events.find(
-      (e) => e.kind === 'tool_call' && e.payload.phase === 'completed',
-    );
-    if (completed?.kind !== 'tool_call' || completed.payload.phase !== 'completed') {
-      throw new Error('unreachable');
-    }
-    expect(completed.payload.truncated).toBe(true);
-
-    // The tool.output ProviderEvent carries the flag through to downstream
-    // consumers (stream-consumer → ToolResultChunk → subagent trace).
-    const toolOutput = events.find((e) => e.type === 'tool.output');
-    if (toolOutput?.type !== 'tool.output') {
-      throw new Error('expected tool.output event');
-    }
-    expect(toolOutput.truncated).toBe(true);
-  });
-
-  it('emits no tool_call events when traceWriter is absent', async () => {
-    const { InMemoryTraceWriter } = await import('../../trace/writer.js');
-    const writer = new InMemoryTraceWriter();
-
-    let callIdx = 0;
-    const streams = [
-      () => fromArray(makeToolUseStream('tu_q', 'search', '{}')),
-      () => fromArray(makeTextStream('done', 'end_turn')),
-    ];
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => streams[callIdx++ % streams.length]!()),
-      },
-    };
-    const dispatcher = makeDispatcher(async () => ({ content: 'r', isError: false }));
-
-    await collect(
-      runTurn({
-        client,
-        messages: [{ role: 'user', content: 'x' }],
-        system: null,
-        tools: [{ name: 'search', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: new AbortController().signal,
-        ctx,
-        // traceWriter omitted intentionally
-      }),
-    );
-    await new Promise((resolve) => setImmediate(resolve));
-    expect(writer.events).toHaveLength(0);
-  });
-});
-
-describe('loop.ts runTurn — render-only diff sidecar', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it('emits tool.diff event when handler returns render.diff', async () => {
-    // First iteration: tool_use. Second iteration: end_turn (to terminate the loop).
-    let callIdx = 0;
-    const client = makeClient(() => {
-      callIdx += 1;
-      if (callIdx === 1) return fromArray(makeToolUseStream('tu_diff', 'edit_file', '{}'));
-      return fromArray(makeTextStream('done', 'end_turn'));
-    });
-
-    const dispatcher = makeDispatcher(() =>
-      Promise.resolve({
-        content: 'Replaced 1 occurrence in /tmp/x.ts',
-        render: {
-          diff: {
-            hunks: [
-              {
-                oldStart: 1,
-                oldLines: 1,
-                newStart: 1,
-                newLines: 1,
-                lines: [{ kind: '-' as const, text: 'a' }, { kind: '+' as const, text: 'b' }],
-              },
-            ],
-            addedLines: 1,
-            removedLines: 1,
-          },
-        },
-      }),
-    );
-
-    const events = await collect(
-      runTurn({
-        client,
-        messages: [{ role: 'user', content: 'edit it' }],
-        system: null,
-        tools: [{ name: 'edit_file', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: new AbortController().signal,
-        ctx,
-      }),
-    );
-
-    // Find tool.output and tool.diff for the same toolUseId.
-    const output = events.find((e) => e.type === 'tool.output');
-    const diff = events.find((e) => e.type === 'tool.diff');
-    expect(output).toBeDefined();
-    expect(diff).toBeDefined();
-    if (output?.type === 'tool.output' && diff?.type === 'tool.diff') {
-      expect(diff.toolUseId).toBe(output.toolUseId);
-      expect(diff.diff.addedLines).toBe(1);
-      expect(diff.diff.removedLines).toBe(1);
-    }
-
-    // Order invariant: tool.diff arrives AFTER tool.output for the same id.
-    const outputIdx = events.findIndex((e) => e.type === 'tool.output');
-    const diffIdx = events.findIndex((e) => e.type === 'tool.diff');
-    expect(diffIdx).toBeGreaterThan(outputIdx);
-  });
-
-  it('does NOT emit tool.diff when handler omits render', async () => {
-    let callIdx = 0;
-    const client = makeClient(() => {
-      callIdx += 1;
-      if (callIdx === 1) return fromArray(makeToolUseStream('tu_no_diff', 'read_file', '{}'));
-      return fromArray(makeTextStream('done', 'end_turn'));
-    });
-    const dispatcher = makeDispatcher(() =>
-      Promise.resolve({ content: 'file contents' }),
-    );
-
-    const events = await collect(
-      runTurn({
-        client,
-        messages: [{ role: 'user', content: 'read it' }],
-        system: null,
-        tools: [{ name: 'read_file', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: new AbortController().signal,
-        ctx,
-      }),
-    );
-
-    expect(events.some((e) => e.type === 'tool.output')).toBe(true);
-    expect(events.some((e) => e.type === 'tool.diff')).toBe(false);
-  });
-
-  it('does NOT include render.diff in the model-facing tool_result block', async () => {
-    // Structural correctness check: the messages array (which is what's
-    // sent back to the model on the next iteration) must contain only the
-    // result.content string — never the diff payload. This is the central
-    // invariant the render channel exists to enforce.
-    let callIdx = 0;
-    const client = makeClient(() => {
-      callIdx += 1;
-      if (callIdx === 1) return fromArray(makeToolUseStream('tu_leak', 'edit_file', '{}'));
-      return fromArray(makeTextStream('done', 'end_turn'));
-    });
-    const dispatcher = makeDispatcher(() =>
-      Promise.resolve({
-        content: 'Replaced 1 occurrence',
-        render: {
-          diff: {
-            hunks: [
-              {
-                oldStart: 1,
-                oldLines: 1,
-                newStart: 1,
-                newLines: 1,
-                lines: [
-                  { kind: '-' as const, text: 'SECRET_BEFORE' },
-                  { kind: '+' as const, text: 'SECRET_AFTER' },
-                ],
-              },
-            ],
-            addedLines: 1,
-            removedLines: 1,
-          },
-        },
-      }),
-    );
-
-    const messages: MessageParam[] = [{ role: 'user', content: 'edit it' }];
-    await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: [{ name: 'edit_file', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: new AbortController().signal,
-        ctx,
-      }),
-    );
-
-    // The loop pushes a `tool_result` MessageParam after dispatching tools.
-    // Walk every tool_result block in the final messages array and assert
-    // that no diff-payload sentinel ever appears in the content the model
-    // would see.
-    const serialized = JSON.stringify(messages);
-    expect(serialized).not.toContain('SECRET_BEFORE');
-    expect(serialized).not.toContain('SECRET_AFTER');
-    expect(serialized).not.toContain('"hunks"');
-    expect(serialized).not.toContain('"addedLines"');
-
-    // And the model's tool_result content should be the one-line summary only.
-    let found = false;
-    for (const msg of messages) {
-      if (msg.role === 'user' && Array.isArray(msg.content)) {
-        for (const block of msg.content) {
-          if (
-            block &&
-            typeof block === 'object' &&
-            'type' in block &&
-            block.type === 'tool_result'
-          ) {
-            found = true;
-            expect((block as { content: unknown }).content).toBe('Replaced 1 occurrence');
-          }
-        }
-      }
-    }
-    expect(found).toBe(true);
-  });
-});
-
-// ─── isTransientServerError ──────────────────────────────────────────────────
-
-describe('isTransientServerError', () => {
-  it('returns true for HTTP 529 (overloaded)', () => {
-    const err = Object.assign(new Error('Overloaded'), { status: 529 });
-    expect(isTransientServerError(err)).toBe(true);
-  });
-
-  it('returns true for HTTP 503 (service unavailable)', () => {
-    const err = Object.assign(new Error('Service Unavailable'), { status: 503 });
-    expect(isTransientServerError(err)).toBe(true);
-  });
-
-  it('returns false for HTTP 429 (rate limit)', () => {
-    const err = Object.assign(new Error('Rate limited'), { status: 429 });
-    expect(isTransientServerError(err)).toBe(false);
-  });
-
-  it('returns false for HTTP 401 (auth)', () => {
-    const err = Object.assign(new Error('Unauthorized'), { status: 401 });
-    expect(isTransientServerError(err)).toBe(false);
-  });
-
-  it('returns false for errors without status', () => {
-    expect(isTransientServerError(new Error('network failure'))).toBe(false);
-  });
-});
-
-// ─── isOverloadedErrorEvent ──────────────────────────────────────────────────
-//
-// Mid-stream overloads carry NO HTTP status (the SDK throws them from inside
-// the stream iterator as `new APIError(undefined, <parsed SSE body>, …)`), so
-// detection must key off the parsed body, not `status`. The body shape from a
-// real Anthropic `event: error` SSE frame is double-nested:
-//   {"type":"error","error":{"type":"overloaded_error","message":"Overloaded"}}
-describe('isOverloadedErrorEvent', () => {
-  it('detects a mid-stream overload from the double-nested SSE body (status undefined)', () => {
-    const err = Object.assign(new Error('Overloaded'), {
-      status: undefined,
-      error: { type: 'error', error: { type: 'overloaded_error', message: 'Overloaded' }, request_id: 'req_x' },
-    });
-    expect(isOverloadedErrorEvent(err)).toBe(true);
-  });
-
-  it('detects a flat overload body shape', () => {
-    const err = Object.assign(new Error('Overloaded'), {
-      error: { type: 'overloaded_error', message: 'Overloaded' },
-    });
-    expect(isOverloadedErrorEvent(err)).toBe(true);
-  });
-
-  it('detects a connection-phase 529/503 by status', () => {
-    expect(isOverloadedErrorEvent(Object.assign(new Error('x'), { status: 529 }))).toBe(true);
-    expect(isOverloadedErrorEvent(Object.assign(new Error('x'), { status: 503 }))).toBe(true);
-  });
-
-  it('returns false for a non-overload error event (e.g. invalid_request)', () => {
-    const err = Object.assign(new Error('bad'), {
-      status: undefined,
-      error: { type: 'error', error: { type: 'invalid_request_error', message: 'bad' } },
-    });
-    expect(isOverloadedErrorEvent(err)).toBe(false);
-  });
-
-  it('returns false for a plain error with no body or status', () => {
-    expect(isOverloadedErrorEvent(new Error('network failure'))).toBe(false);
-    expect(isOverloadedErrorEvent(null)).toBe(false);
-    expect(isOverloadedErrorEvent(undefined)).toBe(false);
-    expect(isOverloadedErrorEvent('overloaded_error')).toBe(false);
-  });
-});
-
-// ─── createWithRetry (via runTurn) ───────────────────────────────────────────
-
-describe('runTurn transient error retry', () => {
-  beforeEach(() => {
-    vi.useFakeTimers();
-  });
-
-  afterEach(() => {
-    vi.useRealTimers();
-  });
-
-  it('retries on 529 and succeeds on subsequent attempt', async () => {
-    let callCount = 0;
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => {
-          callCount++;
-          if (callCount === 1) {
-            throw Object.assign(new Error('Overloaded'), { status: 529, type: 'overloaded_error' });
-          }
-          return fromArray(makeTextStream('recovered'));
-        }),
-      },
-    };
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'ok' }));
-    const messages: MessageParam[] = [{ role: 'user', content: 'hi' }];
-    const abortController = new AbortController();
-
-    const resultPromise = collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: null,
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-      }),
-    );
-
-    // Advance past the retry delay
-    await vi.advanceTimersByTimeAsync(10_000);
-
-    const events = await resultPromise;
-
-    expect(callCount).toBe(2);
-    const errorEvent = events.find((e) => e.type === 'error');
-    expect(errorEvent).toBeUndefined();
-    const completed = events.find((e) => e.type === 'turn.completed');
-    expect(completed).toBeDefined();
-  });
-
-  it('exhausts retries and yields error on persistent 529', async () => {
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => {
-          throw Object.assign(new Error('Overloaded'), { status: 529, type: 'overloaded_error' });
-        }),
-      },
-    };
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'ok' }));
-    const messages: MessageParam[] = [{ role: 'user', content: 'hi' }];
-    const abortController = new AbortController();
-
-    const resultPromise = collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: null,
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-      }),
-    );
-
-    // Advance past all retry delays (5s + 10s + 20s = 35s)
-    await vi.advanceTimersByTimeAsync(40_000);
-
-    const events = await resultPromise;
-
-    // 1 initial + OVERLOAD_MAX_RETRIES retries
-    expect(client.messages.create).toHaveBeenCalledTimes(OVERLOAD_MAX_RETRIES + 1);
-    const errorEvent = events.find((e) => e.type === 'error');
-    expect(errorEvent).toBeDefined();
-    if (errorEvent?.type === 'error') {
-      expect(errorEvent.error.message).toContain('Overloaded');
-    }
-  });
-
-  it('does not retry non-transient errors (e.g. 400)', async () => {
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => {
-          throw Object.assign(new Error('Bad request'), { status: 400 });
-        }),
-      },
-    };
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'ok' }));
-    const messages: MessageParam[] = [{ role: 'user', content: 'hi' }];
-    const abortController = new AbortController();
-
-    const events = await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: null,
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-      }),
-    );
-
-    expect(client.messages.create).toHaveBeenCalledTimes(1);
-    const errorEvent = events.find((e) => e.type === 'error');
-    expect(errorEvent).toBeDefined();
-  });
-
-  it('aborts during retry sleep and yields turn.completed', async () => {
-    let callCount = 0;
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => {
-          callCount++;
-          throw Object.assign(new Error('Overloaded'), { status: 529 });
-        }),
-      },
-    };
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'ok' }));
-    const messages: MessageParam[] = [{ role: 'user', content: 'hi' }];
-    const abortController = new AbortController();
-
-    const resultPromise = collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: null,
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-      }),
-    );
-
-    // Let the first attempt fail, then abort during the sleep
-    await vi.advanceTimersByTimeAsync(100);
-    abortController.abort('interrupted');
-    await vi.advanceTimersByTimeAsync(10_000);
-
-    const events = await resultPromise;
-
-    expect(callCount).toBe(1);
-    const completed = events.find((e) => e.type === 'turn.completed');
-    expect(completed).toBeDefined();
-  });
-});
-
-// ─── Mid-stream overload retry (via runTurn) ─────────────────────────────────
-//
-// Regression for the v3.78.x crash: an Anthropic `overloaded_error` delivered
-// mid-stream (HTTP 200, then an `event: error` SSE frame) is thrown from the
-// SDK stream iterator with `status === undefined`. createWithRetry — status-
-// based and wrapping only the connection-phase messages.create() — never saw
-// it, the auth/usage-limit retry tiers ignored it, and the error event
-// propagated to a fatal turn crash. These tests pin the mid-stream retry path.
-describe('runTurn mid-stream overload retry', () => {
-  beforeEach(() => {
-    vi.useFakeTimers();
-  });
-  afterEach(() => {
-    vi.useRealTimers();
-  });
-
-  // A stream that yields message_start then throws the exact mid-stream
-  // overload shape the SDK produces: status undefined, double-nested
-  // overloaded_error body. translateMessageStream catches the throw and yields
-  // it as an in-band error event — the path runTurn must now retry.
-  function midStreamOverloadStream(): AsyncIterable<RawMessageStreamEvent> {
-    return (async function* () {
-      yield {
-        type: 'message_start',
-        message: {
-          id: 'msg_overload', type: 'message', role: 'assistant', content: [],
-          model: 'claude-test', stop_reason: null, stop_sequence: null, usage: baseUsage(),
-        },
-      } as unknown as RawMessageStreamEvent;
-      throw Object.assign(new Error('Overloaded'), {
-        status: undefined,
-        error: { type: 'error', error: { type: 'overloaded_error', message: 'Overloaded' }, request_id: 'req_overload' },
-      });
-    })();
-  }
-
-  it('retries a mid-stream overload and succeeds on the next attempt', async () => {
-    let callCount = 0;
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => {
-          callCount++;
-          return callCount === 1 ? midStreamOverloadStream() : fromArray(makeTextStream('recovered'));
-        }),
-      },
-    };
-    const resultPromise = collect(
-      runTurn({
-        client, messages: [{ role: 'user', content: 'hi' }], system: null, tools: null,
-        toolDispatcher: makeDispatcher(() => Promise.resolve({ content: 'ok' })),
-        model: 'claude-test', maxTokens: 1024, headers: {}, signal: new AbortController().signal, ctx,
-      }),
-    );
-
-    await vi.advanceTimersByTimeAsync(10_000); // past the first 5s backoff
-    const events = await resultPromise;
-
-    expect(callCount).toBe(2);
-    expect(events.find((e) => e.type === 'error')).toBeUndefined();
-    expect(events.find((e) => e.type === 'turn.completed')).toBeDefined();
-    // The retry emits exactly one stream.retry marker so surfaces can discard
-    // the overloaded attempt's partial text before the recovered re-stream.
-    expect(events.filter((e) => e.type === 'stream.retry')).toHaveLength(1);
-  });
-
-  it('exhausts the retry budget on a persistent mid-stream overload and yields the error', async () => {
-    const client: AnthropicClientLike = {
-      messages: { create: vi.fn(() => midStreamOverloadStream()) },
-    };
-    const resultPromise = collect(
-      runTurn({
-        client, messages: [{ role: 'user', content: 'hi' }], system: null, tools: null,
-        toolDispatcher: makeDispatcher(() => Promise.resolve({ content: 'ok' })),
-        model: 'claude-test', maxTokens: 1024, headers: {}, signal: new AbortController().signal, ctx,
-      }),
-    );
-
-    await vi.advanceTimersByTimeAsync(40_000); // past all backoffs (5s + 10s + 20s)
-    const events = await resultPromise;
-
-    expect(client.messages.create).toHaveBeenCalledTimes(OVERLOAD_MAX_RETRIES + 1);
-    // One stream.retry per backoff/re-drive — OVERLOAD_MAX_RETRIES total (the
-    // final, exhausted attempt yields the error, not another retry marker).
-    expect(events.filter((e) => e.type === 'stream.retry')).toHaveLength(OVERLOAD_MAX_RETRIES);
-    const errorEvent = events.find((e) => e.type === 'error');
-    expect(errorEvent).toBeDefined();
-    if (errorEvent?.type === 'error') {
-      expect(errorEvent.error.message).toContain('Overloaded');
-    }
-  });
-
-  it('does NOT retry a non-overload mid-stream error', async () => {
-    let callCount = 0;
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => {
-          callCount++;
-          return (async function* () {
-            yield {
-              type: 'message_start',
-              message: {
-                id: 'msg_bad', type: 'message', role: 'assistant', content: [],
-                model: 'claude-test', stop_reason: null, stop_sequence: null, usage: baseUsage(),
-              },
-            } as unknown as RawMessageStreamEvent;
-            throw Object.assign(new Error('Invalid request'), {
-              status: undefined,
-              error: { type: 'error', error: { type: 'invalid_request_error', message: 'Invalid request' } },
-            });
-          })();
-        }),
-      },
-    };
-    const events = await collect(
-      runTurn({
-        client, messages: [{ role: 'user', content: 'hi' }], system: null, tools: null,
-        toolDispatcher: makeDispatcher(() => Promise.resolve({ content: 'ok' })),
-        model: 'claude-test', maxTokens: 1024, headers: {}, signal: new AbortController().signal, ctx,
-      }),
-    );
-
-    expect(callCount).toBe(1); // no backoff, no retry — surfaces immediately
-    expect(events.find((e) => e.type === 'error')).toBeDefined();
-    // A non-overload error surfaces immediately — no retry, so no marker.
-    expect(events.find((e) => e.type === 'stream.retry')).toBeUndefined();
-  });
-
-  it('aborts during the mid-stream retry backoff and yields turn.completed', async () => {
-    let callCount = 0;
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn(() => {
-          callCount++;
-          return midStreamOverloadStream();
-        }),
-      },
-    };
-    const abortController = new AbortController();
-    const resultPromise = collect(
-      runTurn({
-        client, messages: [{ role: 'user', content: 'hi' }], system: null, tools: null,
-        toolDispatcher: makeDispatcher(() => Promise.resolve({ content: 'ok' })),
-        model: 'claude-test', maxTokens: 1024, headers: {}, signal: abortController.signal, ctx,
-      }),
-    );
-
-    await vi.advanceTimersByTimeAsync(100); // first attempt overloads, enters backoff
-    abortController.abort('interrupted');
-    await vi.advanceTimersByTimeAsync(10_000);
-    const events = await resultPromise;
-
-    expect(callCount).toBe(1); // aborted before the retry attempt
-    expect(events.find((e) => e.type === 'turn.completed')).toBeDefined();
-  });
-});
-
-// ─── Regression: orphan tool_use prevention ────────────────────────────────────
-//
-// Anthropic Messages API rejects any user-turn submit whose prior assistant
-// message contains a `tool_use` block not immediately followed by a user
-// message of matching `tool_result` blocks:
-//
-//   400 messages.N: `tool_use` ids were found without `tool_result` blocks
-//   immediately after: toolu_XXX
-//
-// runTurn must never leave history in that shape. These tests pin the two
-// codepaths that historically did:
-//
-//   1. stopReason != 'tool_use' but assistantBlocks contains a tool_use block
-//      (e.g. `max_tokens` truncated the message after the tool_use opened).
-//      Pre-fix: the tool_use block was pushed verbatim — instant orphan.
-//      Post-fix: tool_use blocks are filtered from the pushed assistant turn.
-//
-//   2. stopReason == 'tool_use' but the dispatcher throws synchronously
-//      OUTSIDE the dispatcher's own try/catch (e.g. accessor on
-//      `input.toolDispatcher` throws). Pre-fix: assistant tool_use push
-//      remained in history. Post-fix: rollback splices it back out.
-
-describe('loop.ts runTurn — orphan tool_use prevention', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it('strips tool_use blocks when stopReason is not tool_use (max_tokens truncation)', async () => {
-    // Stream: a tool_use block opens AND closes, but message_delta carries
-    // stop_reason='max_tokens' (model ran out of budget before the model would
-    // have actually requested a dispatch). assistantBlocks contains the
-    // tool_use block; runTurn must NOT leak it into history.
-    const truncatedToolUseStream: RawMessageStreamEvent[] = [
-      {
-        type: 'message_start',
-        message: {
-          id: 'msg_truncate',
-          type: 'message',
-          role: 'assistant',
-          content: [],
-          model: 'claude-test',
-          stop_reason: null,
-          stop_sequence: null,
-          usage: baseUsage(),
-        },
-      } as unknown as RawMessageStreamEvent,
-      {
-        type: 'content_block_start',
-        index: 0,
-        content_block: { type: 'tool_use', id: 'toolu_orphan', name: 'read_file', input: {} },
-      } as unknown as RawMessageStreamEvent,
-      {
-        type: 'content_block_delta',
-        index: 0,
-        delta: { type: 'input_json_delta', partial_json: '{"file":"a.ts"}' },
-      } as unknown as RawMessageStreamEvent,
-      {
-        type: 'content_block_stop',
-        index: 0,
-      } as unknown as RawMessageStreamEvent,
-      {
-        type: 'message_delta',
-        delta: { stop_reason: 'max_tokens', stop_sequence: null },
-        usage: { output_tokens: 4 },
-      } as unknown as RawMessageStreamEvent,
-      { type: 'message_stop' } as unknown as RawMessageStreamEvent,
-    ];
-
-    const client = makeClient(() => fromArray(truncatedToolUseStream));
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'never called' }));
-    const messages: MessageParam[] = [{ role: 'user', content: 'do a thing' }];
-    const abortController = new AbortController();
-
-    await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: null,
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-      }),
-    );
-
-    // History invariant: no message ends in an unmatched tool_use block.
-    // Equivalent: no assistant message anywhere carries tool_use without a
-    // following user message of tool_result blocks. With this stream we
-    // expect the assistant turn to be either skipped entirely (no usable
-    // content) or to contain only non-tool_use blocks. The tool_use must
-    // NOT appear in history.
-    const flatTooluses: string[] = [];
-    for (const m of messages) {
-      if (typeof m.content === 'string') continue;
-      for (const b of m.content as ContentBlockParam[]) {
-        if (b.type === 'tool_use') flatTooluses.push((b as { id: string }).id);
-      }
-    }
-    expect(flatTooluses).not.toContain('toolu_orphan');
-  });
-
-  it('rolls back the assistant tool_use push when the dispatcher accessor throws', async () => {
-    // Build a dispatcher whose `executeBatch` is a getter that throws on
-    // access. This simulates a throw between "tool_use is committed to
-    // history" and "tool_result is committed to history" that NOT absorbed
-    // by the existing executeBatch try/catch (which only catches when the
-    // function call itself throws, after access succeeds).
-    const dispatcher: ToolDispatcherLike = {
-      execute: vi.fn(() => Promise.reject(new Error('should use batch'))),
-    };
-    Object.defineProperty(dispatcher, 'executeBatch', {
-      get() {
-        throw new Error('synthetic mid-turn failure');
-      },
-    });
-
-    const stream = makeToolUseStream('toolu_will_orphan', 'read_file', '{"file":"a.ts"}');
-    const client = makeClient(() => fromArray(stream));
-    const messages: MessageParam[] = [{ role: 'user', content: 'do it' }];
-    const initialLength = messages.length;
-    const abortController = new AbortController();
-
-    let caught: Error | null = null;
-    try {
-      await collect(
-        runTurn({
-          client,
-          messages,
-          system: null,
-          tools: null,
-          toolDispatcher: dispatcher,
-          model: 'claude-test',
-          maxTokens: 1024,
-          headers: {},
-          signal: abortController.signal,
-          ctx,
-        }),
-      );
-    } catch (err) {
-      caught = err as Error;
-    }
-
-    // The throw is intentional — runTurn re-throws after rolling back. The
-    // invariant under test is: history does NOT carry the orphan tool_use.
-    expect(caught).toBeInstanceOf(Error);
-    expect(messages.length).toBe(initialLength);
-    for (const m of messages) {
-      if (typeof m.content === 'string') continue;
-      for (const b of m.content as ContentBlockParam[]) {
-        expect(b.type).not.toBe('tool_use');
-      }
-    }
-  });
-
-  it('leaves history matched (tool_use + tool_result) on the normal dispatch path', async () => {
-    // Sanity guard: the rollback wrapper must not corrupt the happy path.
-    // After one tool_use round + dispatch, history should contain exactly
-    // the assistant tool_use turn followed by the user tool_result turn.
-    const stream = makeToolUseStream('toolu_ok', 'read_file', '{"file":"a.ts"}');
-    const client = makeClient(() =>
-      // After the dispatch turn, the model would emit a follow-up — but we
-      // only need one round, so the second call gets an end_turn text stream.
-      fromArray(stream),
-    );
-    // Replace the create mock with a sequence: first call returns the tool_use
-    // stream, second call returns an end_turn text stream.
-    let callIdx = 0;
-    client.messages.create = vi.fn(() => {
-      callIdx++;
-      return callIdx === 1
-        ? fromArray(stream)
-        : fromArray(makeTextStream('done', 'end_turn'));
-    });
-
-    const dispatcher = makeBatchDispatcher(async (calls) =>
-      calls.map(() => ({ content: 'file contents', isError: false as const })),
-    );
-    const messages: MessageParam[] = [{ role: 'user', content: 'read it' }];
-    const abortController = new AbortController();
-
-    await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: null,
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-      }),
-    );
-
-    // Find the assistant tool_use message and verify the very next message
-    // is a user message whose content covers the same tool_use_id.
-    let toolUseIdx = -1;
-    for (let i = 0; i < messages.length; i++) {
-      const m = messages[i]!;
-      if (m.role !== 'assistant' || typeof m.content === 'string') continue;
-      for (const b of m.content as ContentBlockParam[]) {
-        if (b.type === 'tool_use') {
-          toolUseIdx = i;
-          break;
-        }
-      }
-      if (toolUseIdx >= 0) break;
-    }
-    expect(toolUseIdx).toBeGreaterThanOrEqual(0);
-    const next = messages[toolUseIdx + 1];
-    expect(next?.role).toBe('user');
-    expect(typeof next?.content).not.toBe('string');
-    const nextBlocks = next!.content as ContentBlockParam[];
-    expect(nextBlocks.some((b) => b.type === 'tool_result' && (b as { tool_use_id: string }).tool_use_id === 'toolu_ok')).toBe(true);
-  });
-
-  // Regression: the Anthropic Messages API rejects extra fields on custom
-  // tool definitions with 400 `tools.0.custom.category: Extra inputs are
-  // not permitted`. AnthropicToolDef carries internal classification
-  // metadata (`category`, `concurrencySafe`, `riskClass`) that must be
-  // stripped before send. `toWireTool` projects to the wire-safe shape;
-  // this test pins the contract.
-  it('strips internal classification metadata (category, concurrencySafe, riskClass) before sending tools to the API', async () => {
-    let captured: { tools?: unknown } | undefined;
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn((params: unknown) => {
-          captured = params as { tools?: unknown };
-          return fromArray(makeTextStream('hi'));
-        }),
-      },
-    };
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'ok' }));
-
-    const messages: MessageParam[] = [{ role: 'user', content: 'hi' }];
-    const abortController = new AbortController();
-
-    await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: [
-          {
-            name: 'bash',
-            description: 'run a shell command',
-            input_schema: { type: 'object', properties: { command: { type: 'string' } }, required: ['command'] },
-            category: 'shell',
-            concurrencySafe: false,
-            riskClass: 'destructive',
-          },
-        ],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-      }),
-    );
-
-    expect(captured).toBeDefined();
-    const sentTools = (captured!.tools as Array<Record<string, unknown>> | undefined) ?? [];
-    expect(sentTools).toHaveLength(1);
-    const sent = sentTools[0]!;
-    // wire-safe fields preserved
-    expect(sent['name']).toBe('bash');
-    expect(sent['description']).toBe('run a shell command');
-    expect(sent['input_schema']).toBeDefined();
-    // internal metadata MUST be stripped — these caused the 400
-    expect(sent).not.toHaveProperty('category');
-    expect(sent).not.toHaveProperty('concurrencySafe');
-    expect(sent).not.toHaveProperty('riskClass');
-    // exhaustive: only the three wire-safe keys leave the boundary
-    expect(Object.keys(sent).sort()).toEqual(['description', 'input_schema', 'name']);
-  });
-
-  it('omits `description` when not provided (does not emit `description: undefined`)', async () => {
-    let captured: { tools?: unknown } | undefined;
-    const client: AnthropicClientLike = {
-      messages: {
-        create: vi.fn((params: unknown) => {
-          captured = params as { tools?: unknown };
-          return fromArray(makeTextStream('hi'));
-        }),
-      },
-    };
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'ok' }));
-
-    await collect(
-      runTurn({
-        client,
-        messages: [{ role: 'user', content: 'hi' }],
-        system: null,
-        tools: [{ name: 'no_desc', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: new AbortController().signal,
-        ctx,
-      }),
-    );
-
-    const sentTools = (captured!.tools as Array<Record<string, unknown>>);
-    expect(Object.keys(sentTools[0]!).sort()).toEqual(['input_schema', 'name']);
-  });
-});
-
-// ─── onUsageProgress mid-turn hook ───────────────────────────────────────────
-
-describe('loop.ts runTurn — onUsageProgress mid-turn hook', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it('fires onUsageProgress once per round with monotonically non-decreasing cumulative tokens', async () => {
-    // Round 1: tool_use response with input=100, output=20 (via message_start + message_delta).
-    // Round 2: final text response with input=150, output=30.
-    // Streams use the same structure as makeToolUseStream / makeTextStream but
-    // with explicit token counts so the assertion is deterministic.
-    function makeToolUseStreamWithUsage(
-      toolId: string,
-      toolName: string,
-      inputJson: string,
-      inputTokens: number,
-      outputTokens: number,
-    ): RawMessageStreamEvent[] {
-      return [
-        {
-          type: 'message_start',
-          message: {
-            id: 'msg_progress_t',
-            type: 'message',
-            role: 'assistant',
-            content: [],
-            model: 'claude-test',
-            stop_reason: null,
-            stop_sequence: null,
-            usage: {
-              input_tokens: inputTokens,
-              output_tokens: 0,
-              cache_creation_input_tokens: 0,
-              cache_read_input_tokens: 0,
-            },
-          },
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'content_block_start',
-          index: 0,
-          content_block: { type: 'tool_use', id: toolId, name: toolName, input: {} },
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'content_block_delta',
-          index: 0,
-          delta: { type: 'input_json_delta', partial_json: inputJson },
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'content_block_stop',
-          index: 0,
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'message_delta',
-          delta: { stop_reason: 'tool_use', stop_sequence: null },
-          usage: { output_tokens: outputTokens },
-        } as unknown as RawMessageStreamEvent,
-        { type: 'message_stop' } as unknown as RawMessageStreamEvent,
-      ];
-    }
-
-    function makeTextStreamWithUsage(
-      text: string,
-      inputTokens: number,
-      outputTokens: number,
-    ): RawMessageStreamEvent[] {
-      return [
-        {
-          type: 'message_start',
-          message: {
-            id: 'msg_progress_txt',
-            type: 'message',
-            role: 'assistant',
-            content: [],
-            model: 'claude-test',
-            stop_reason: null,
-            stop_sequence: null,
-            usage: {
-              input_tokens: inputTokens,
-              output_tokens: 0,
-              cache_creation_input_tokens: 0,
-              cache_read_input_tokens: 0,
-            },
-          },
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'content_block_start',
-          index: 0,
-          content_block: { type: 'text', text: '' },
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'content_block_delta',
-          index: 0,
-          delta: { type: 'text_delta', text },
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'content_block_stop',
-          index: 0,
-        } as unknown as RawMessageStreamEvent,
-        {
-          type: 'message_delta',
-          delta: { stop_reason: 'end_turn', stop_sequence: null },
-          usage: { output_tokens: outputTokens },
-        } as unknown as RawMessageStreamEvent,
-        { type: 'message_stop' } as unknown as RawMessageStreamEvent,
-      ];
-    }
-
-    let callIdx = 0;
-    const client = makeClient(() => {
-      callIdx += 1;
-      if (callIdx === 1) {
-        // Round 1: tool_use with input=100, output=20
-        return fromArray(makeToolUseStreamWithUsage('toolu_prog', 'read_file', '{}', 100, 20));
-      }
-      // Round 2: final text with input=150, output=30
-      return fromArray(makeTextStreamWithUsage('Done', 150, 30));
-    });
-
-    const onUsageProgress = vi.fn();
-    const dispatcher = makeDispatcher(() => Promise.resolve({ content: 'ok' }));
-    const messages: MessageParam[] = [{ role: 'user', content: 'progress test' }];
-    const abortController = new AbortController();
-
-    await collect(
-      runTurn({
-        client,
-        messages,
-        system: null,
-        tools: [{ name: 'read_file', input_schema: { type: 'object' } }],
-        toolDispatcher: dispatcher,
-        model: 'claude-test',
-        maxTokens: 1024,
-        headers: {},
-        signal: abortController.signal,
-        ctx,
-        onUsageProgress,
-      }),
-    );
-
-    // Called at least once per round (round 1 = tool_use, round 2 = end_turn).
-    expect(onUsageProgress).toHaveBeenCalledTimes(2);
-
-    // Extract the full usage objects from each call for exact-value checks.
-    const call0 = onUsageProgress.mock.calls[0]?.[0] as {
-      inputTokens?: number;
-      outputTokens?: number;
-    };
-    const call1 = onUsageProgress.mock.calls[1]?.[0] as {
-      inputTokens?: number;
-      outputTokens?: number;
-    };
-
-    // Exact-value anchors — these prove per-round firing, not just that the
-    // hook was called:
-    //
-    // Round 1: translateMessageStream merges message_start.usage
-    //   { input_tokens: 100, output_tokens: 0 } with message_delta.usage
-    //   { output_tokens: 20 } → { input_tokens: 100, output_tokens: 20 }.
-    //   sumProviderUsage({}, round1) → cumulative = { inputTokens: 100,
-    //   outputTokens: 20 }. onUsageProgress receives exactly this.
-    expect(call0?.inputTokens).toBe(100);
-    expect(call0?.outputTokens).toBe(20);
-
-    // Round 2: input_tokens: 150, output_tokens: 30.
-    //   sumProviderUsage(round1, round2) → { inputTokens: 250, outputTokens: 50 }.
-    //   onUsageProgress receives the NEW cumulative (not round-2's delta alone).
-    expect(call1?.inputTokens).toBe(250);
-    expect(call1?.outputTokens).toBe(50);
-
-    // Monotonicity is now a corollary of the exact values, but keep it as an
-    // explicit guard for future token-count changes that might regress ordering.
-    const tokenTotals = [call0, call1].map((u) =>
-      (u?.inputTokens ?? 0) + (u?.outputTokens ?? 0),
-    );
-    for (let i = 1; i < tokenTotals.length; i++) {
-      expect(tokenTotals[i]).toBeGreaterThanOrEqual(tokenTotals[i - 1]!);
     }
   });
 });
